requirements.txt
Flask
requests
gunicorn
mysql-connector-python
python-dotenv
APScheduler

